requirements.txt
asgiref==3.5.0
async-generator==1.10
attrs==22.1.0
autopep8==2.0.0
beautifulsoup4==4.11.1
boto3==1.22.10
botocore==1.25.10
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
distlib==0.3.4
Django==4.0.4
django-cors-headers==3.13.0
django-storages==1.12.3
djangorestframework==3.14.0
feedparser==6.0.10
filelock==3.6.0
gunicorn==20.1.0
h11==0.13.0
idna==3.4
jmespath==1.0.0
mysql-connector-python==8.0.31
mysqlclient==2.1.0
outcome==1.2.0
pandas==1.4.4
Pillow==9.1.0
platformdirs==2.5.2
pprintpp==0.4.0
protobuf==3.20.1
psycopg2-binary==2.9.5
pycodestyle==2.9.1
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
qrcode==6.1
requests==2.28.1
s3transfer==0.5.2
selenium==4.4.3
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tomli==2.0.1
tqdm==4.64.1
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.1
urllib3==1.26.9
virtualenv==20.14.1
whitenoise==6.0.0
wsproto==1.2.0
